updated return type of save word added parameter category to save label

+++ src/main/java/com/datapath/procurementdata/documentmatcher/dao/repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.datapath.procurementdata.documentmatcher.dao.repository;
 import com.datapath.procurementdata.documentmatcher.dao.entity.CategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<CategoryEntity, Long> {
+public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
 }
